pyproject.toml
[tool.poetry]
name = "lit-lambdas"
version = "0.1.0"
description = ""
authors = ["Uriel Mandujano <uriel.mandujano14@gmail.com>"]
packages = [
    { include = "lit_lambdas", from = "lit_lambdas" },
]

[tool.poetry.dependencies]
python = "^3.8"
httpx = "^0.19.0"
structlog = "^21.1.0"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
black = "^21.6b0"
mypy = "^0.902"
isort = "^5.8.0"
"aws-cdk.aws-lambda" = "^1.108.1"
"aws-cdk.core" = "^1.108.1"
"aws-cdk.aws-lambda-python" = "^1.119.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
src_paths = ["lit_lambdas", "test", "cdk"]

[tool.mypy]
python_version = "3.8"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests"
]